+++ infrastructure/src/main/java/com/ros/aop/aspect/AddBookAspect.java
@@ -1,7 +1,7 @@
 package com.ros.aop.aspect;
 
-import com.ros.aop.audit_entity.CustomLog;
-import com.ros.aop.audit_service.service.BookAuditService;
+import com.ros.aop.audit_repository.AuditDAO;
+import com.ros.aop.audit_repository.CustomLog;
 import org.aspectj.lang.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -11,11 +11,11 @@ import org.springframework.stereotype.Component;
 @Component
 public class AddBookAspect {
 
-    private final BookAuditService bookAuditService;
+    private final AuditDAO auditDAO;
 
     @Autowired
-    public AddBookAspect(@Qualifier("bookAuditServiceImpl") BookAuditService  bookAuditService){
-        this.bookAuditService = bookAuditService;
+    public AddBookAspect(@Qualifier("auditDAOJdbcImpl") AuditDAO auditDAO){
+        this.auditDAO = auditDAO;
     }
 
     @Pointcut("execution(public void com.ros.inbound.controllers.BookController.addBook(..))")
@@ -24,13 +24,13 @@ public class AddBookAspect {
     @AfterReturning("forAddBookMethod()")
     public void afterReturningAddBookAdvice(){
         CustomLog log = new CustomLog( null, "NEW BOOK WAS ADDED");
-        bookAuditService.logAddBookAfter(log);
+        auditDAO.createLog(log);
     }
 
     @AfterThrowing("forAddBookMethod()")
     public void afterThrowingAddBookAdvice(){
         CustomLog log = new CustomLog(null,"ERROR");
-        bookAuditService.logAddBookAfter(log);
+        auditDAO.createLog(log);
     }
 
 }
