+++ infrastructure/src/main/java/com/ros/aop/audit_repository/AuditDAO.java
@@ -1,0 +1,5 @@
+package com.ros.aop.audit_repository;
+
+public interface AuditDAO {
+    void createLog(CustomLog log);
+}
