+++ infrastructure/src/main/java/com/ros/aop/audit_repository/AuditDAOJdbcImpl.java
@@ -1,7 +1,5 @@
-package com.ros.aop.audit_repository.daoImpl;
+package com.ros.aop.audit_repository;
 
-import com.ros.aop.audit_entity.CustomLog;
-import com.ros.aop.audit_repository.dao.AuditDAO;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
@@ -14,7 +12,6 @@ public class AuditDAOJdbcImpl implements AuditDAO {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-
     @Override
     public void createLog(CustomLog log) {
         String procedureCall = "{call GENERATE_LOG(?, ?)}";
@@ -23,6 +20,5 @@ public class AuditDAOJdbcImpl implements AuditDAO {
                 log.description(),
                 log.actionType()
         );
-        System.out.println(log);
     }
 }
