+++ infrastructure/src/main/java/com/ros/aop/audit_repository/CustomLog.java
@@ -1,4 +1,4 @@
-package com.ros.aop.audit_entity;
+package com.ros.aop.audit_repository;
 
 public record CustomLog(String description, String actionType) {
