--- infrastructure/src/main/java/com/ros/aop/audit_repository/dao/AuditDAO.java
@@ -1,7 +1,0 @@
-package com.ros.aop.audit_repository.dao;
-
-import com.ros.aop.audit_entity.CustomLog;
-
-public interface AuditDAO {
-    void createLog(CustomLog log);
-}
