--- infrastructure/src/main/java/com/ros/aop/audit_service/service/BookAuditService.java
@@ -1,7 +1,0 @@
-package com.ros.aop.audit_service.service;
-
-import com.ros.aop.audit_entity.CustomLog;
-
-public interface BookAuditService {
-    void logAddBookAfter(CustomLog log);
-}
